// src/app/features/mastertables/tblPartyMaster/models/tblPartyMaster-Update.model.ts
export interface TblPartyMasterUpdate {
	fldId: number;
	fldFKPartyCategory: number;
	fldCode: string;
	fldName: string;
	fldShortName: string;
	fldAddress1: string;
	fldAddress2: string;
	fldFKArea: number;
	fldState: string;
	fldCity: string;
	fldArea: string;
	fldPincode: string;
	fldStateCode: string;
	fldPANNo: string;
	fldGSTNo: string;
	fldContactPersonName: string;
	fldFKDesignation: number;
	fldMobileNo: string;
	fldMailId: string;
	fldImagePath: string;
	fldIsActive: boolean;
	fldCreatedBy: number;
	fldCreatedDt: Date;
	fldModifiedBy: number;
	fldModifiedDt: Date;
}